Feature/review

+++ backend/src/api/reviews/review.model.ts
@@ -5,6 +5,8 @@ import { Review, Prisma, Image } from '@prisma/client';
 import prisma from '../../config/database';
 import { z } from 'zod';
 import { AppError } from '../../middleware/errorHandler';
+import { CustomMulterFile } from './review.controller';
+import S3Service from '../../utils/s3Service';
 
 // zodライブラリを使用してプロパティの型や制約を定義
 export const reviewSchema = z.object({
@@ -21,7 +23,7 @@ export const reviewSchema = z.object({
     .transform((str) => new Date(str))
     .optional(),
   content: z.string().min(1).max(2000),
-  image: z
+  images: z
     .array(
       z.object({
         imageUrl: z.string().url(),
@@ -36,7 +38,7 @@ export const reviewSchema = z.object({
 export type ReviewInput = z.infer<typeof reviewSchema>;
 
 // reviewとimageを組み合わせた型
-export type ReviewWithImages = Review & { image: Image[] };
+export type ReviewWithImages = Review & { images?: Image[] };
 
 // formatToJapanTime が文字列で返される為、型整合性を合わせる
 export type ReviewJapanTime = Omit<Review, 'createdAt' | 'updatedAt' | 'purchaseDate'> & {
@@ -45,29 +47,59 @@ export type ReviewJapanTime = Omit<Review, 'createdAt' | 'updatedAt' | 'purchase
   purchaseDate: string | null;
 };
 
+// imageの型ガード関数
+const isReviewWithImages = (review: ReviewModel | AppError | undefined): review is ReviewModel => {
+  return review !== undefined && !(review instanceof AppError) && 'images' in review;
+};
+
+// 画像処理関数
+export const processImages = async (files: CustomMulterFile[]) => {
+  if (files && files.length > 0) {
+    const imageUrls = await Promise.all(files.map(S3Service.uploadToS3));
+    // indexを使用し、メイン画像を０とする
+    return files.map((file, index) => ({
+      imageUrl: imageUrls[index],
+      order: file.order,
+      isMain: index === 0,
+    }));
+  }
+  return [];
+};
+
 export class ReviewModel {
+  raw: ReviewWithImages;
+  images: Image[];
+
+  constructor(params: ReviewWithImages) {
+    this.raw = params;
+    this.images = params.images || [];
+  }
   // レビュー検索
-  getReviewById = async (id: number): Promise<ReviewWithImages | AppError | undefined> => {
+  static findById = async (id: number): Promise<ReviewModel | undefined> => {
     if (isNaN(id)) {
-      return new AppError('Invalid ID', 400);
+      throw new Error('Invalid ID');
     }
     const review = await prisma.review.findUnique({
       where: { id },
-      include: { image: true },
+      include: { images: true },
     });
     if (review) {
-      return review;
+      return new ReviewModel(review);
     }
+
     return undefined;
   };
-
   // レビューの生成
-  createReview = async (reviewData: ReviewInput): Promise<ReviewWithImages | AppError> => {
+  static createReview = async (reviewData: ReviewInput, files: CustomMulterFile[]): Promise<ReviewWithImages> => {
+    // zodスキーマでバリデーション実施
     const parseReview = await reviewSchema.safeParse(reviewData);
     if (!parseReview.success) {
-      return new AppError(parseReview.error.message, 400);
+      throw new Error('validation Error');
     }
-    const { image, ...reviewDataWithoutImage } = parseReview.data;
+    const { images, ...reviewDataWithoutImage } = parseReview.data;
+    // リクエストの画像ファイルを専用関数でS3へアップロード
+    const processedImages = await processImages(files);
+
     return await prisma.review.create({
       data: {
         ...reviewDataWithoutImage,
@@ -77,24 +109,35 @@ export class ReviewModel {
             : null,
         // ISO-8601形式の日時をDateオブジェクトへ変換
         purchaseDate: reviewDataWithoutImage.purchaseDate ? new Date(reviewDataWithoutImage.purchaseDate) : null,
-        image: {
-          create: image,
+        images: {
+          create: processedImages.map((image) => ({
+            imageUrl: image.imageUrl,
+            order: image.order,
+          })),
         },
       },
-      include: { image: true },
+      include: { images: true },
     });
   };
 
   // レビュー更新
-  updateReview = async (id: number, reviewData: ReviewInput): Promise<ReviewWithImages | AppError> => {
+  static updateReview = async (
+    id: number,
+    reviewData: ReviewInput,
+    files: CustomMulterFile[],
+  ): Promise<ReviewModel> => {
     if (isNaN(id)) {
-      return new AppError('Invalid ID', 400);
+      throw new Error('Invalid ID');
     }
     const parseReview = await reviewSchema.safeParse(reviewData);
     if (!parseReview.success) {
-      throw new AppError(parseReview.error.message, 400);
+      throw new Error(parseReview.error.message);
     }
-    const { image, ...reviewDataWithoutImage } = parseReview.data;
+    const { images, ...reviewDataWithoutImage } = parseReview.data;
+
+    // リクエストの画像ファイルを専用関数でS3へアップロード
+    const processedImages = await processImages(files);
+
     const review = await prisma.review.update({
       where: { id },
       data: {
@@ -105,26 +148,39 @@ export class ReviewModel {
             : null,
         // ISO-8601形式の日時をDateオブジェクトへ変換
         purchaseDate: reviewDataWithoutImage.purchaseDate ? new Date(reviewDataWithoutImage.purchaseDate) : null,
-        image: {
+        images: {
           deleteMany: {},
-          create: image,
+          create: images,
         },
       },
-      include: { image: true },
+      include: { images: true },
     });
-    return review;
+    return new ReviewModel(review);
   };
 
   // レビュー削除
-  deleteReview = async (id: number) => {
+  static deleteReview = async (id: number) => {
     if (isNaN(id)) {
-      return new AppError('Invalid ID', 400);
+      throw new Error('Invalid ID');
     }
+    // S3の画像データ削除の為レビュー内容を取得
+    const review = await this.findById(id);
+    if (!review) {
+      throw new Error('Review not found');
+    }
+
+    if (isReviewWithImages(review) && review.images && review.images.length > 0) {
+      // 逐次処理で１個ずつ画像削除処理を行う
+      for (const image of review.images) {
+        // レビューのimageからurlを取得しdeleteS3Objectへ渡す
+        await S3Service.deleteS3Object(image.imageUrl);
+      }
+    }
+
     return await prisma.review.delete({
       where: { id },
-      include: { image: true },
     });
   };
 }
 
-export default new ReviewModel();
+export default ReviewModel;
